Pull out each section of the service schema to their own file
